# swagger.yaml
openapi: 3.1.0
info:
  title: GVHMR API
  version: 1.0.0
  description: Tài liệu API cho hệ thống phân tích và xử lý video GVHMR
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
  /api/v1/health:
    get:
      summary: Kiểm tra sức khỏe (Health Check)
      description: API dùng để kiểm tra trạng thái hoạt động của server và GPU.
      operationId: health_check
      responses:
        '200':
          description: Trạng thái server hoạt động bình thường
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  gpu_available:
                    type: boolean
                    example: true
                  device_name:
                    type: string
                    example: "NVIDIA GeForce RTX 4090"

  /api/v1/video/upload:
    post:
      summary: Upload Video
      description: API tải video từ client lên server để chuẩn bị cho quá trình xử lý. Trả về `video_id` dùng cho bước tạo job.
      operationId: upload_video
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: string
                  format: binary
                  description: File video cần xử lý (VD .mp4)
      responses:
        '200':
          description: Upload thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  video_id:
                    type: string
                    example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                  filename:
                    type: string
                    example: "athlete_run.mp4"
                  saved_path:
                    type: string
        '500':
          description: Lỗi hệ thống trong quá trình lưu file

  /api/v1/jobs:
    post:
      summary: Tạo tác vụ xử lý (Create Job)
      description: Tạo một tác vụ chạy ngầm để phân tích video đã upload dựa vào `video_id`.
      operationId: create_job
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
      responses:
        '200':
          description: Tác vụ đã được tạo thành công và đang chạy ngầm
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Da bat dau tac vu xu ly tac vu ngam"
                  job_id:
                    type: string
                    example: "550e8400-e29b-41d4-a716-446655440000"
        '404':
          description: Không tìm thấy video tương ứng với `video_id` cung cấp
        '503':
          description: Server chưa tải xong Model vào GPU

  /api/v1/jobs/{job_id}:
    get:
      summary: Kiểm tra trạng thái tác vụ (Get Job Status)
      description: Trả về trạng thái tiến độ hiện tại của quá trình xử lý video, bao gồm URL kết quả nếu đã hoàn thành.
      operationId: get_job_status
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: job_id
          required: true
          description: ID của tác vụ cần kiểm tra tiến độ
          schema:
            type: string
      responses:
        '200':
          description: Trạng thái hiện tại của tác vụ
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [PENDING, PROCESSING, COMPLETED, FAILED]
                    example: "COMPLETED"
                  progress:
                    type: string
                    example: "Hoàn thành"
                  created_at:
                    type: string
                    format: date-time
                    example: "2026-08-14T04:15:30Z"
                  result:
                    type: object
                    nullable: true
                    properties:
                      output_video_url:
                        type: string
                        example: "/api/v1/download/incam_global_horiz_video.mp4"
                      result_file_path:
                        type: string
                        example: "output/result/hmr4d_results.pt"
        '404':
          description: Không tìm thấy tác vụ với `job_id` cung cấp

  /api/v1/jobs/{job_id}/download:
    get:
      summary: Tải kết quả tác vụ (Download Job Results)
      description: Nén (ZIP) file video và file tọa độ 3D (.pt) của tác vụ rồi tải về máy tính.
      operationId: download_job_results
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: job_id
          required: true
          description: ID của tác vụ cần tải kết quả
          schema:
            type: string
      responses:
        '200':
          description: Tải xuống thành công file ZIP
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          description: Tác vụ chưa hoàn thành
        '404':
          description: Không tìm thấy tác vụ hoặc file kết quả không tồn tại
        '500':
          description: Lỗi thiếu thông tin file kết quả

  /api/v1/metrics/evaluate:
    post:
      summary: Tính toán sai số MPJPE và PA-MPJPE (Calculate Evaluation Metrics)
      description: Tính toán sai số trung bình vị trí khớp MPJPE (Pelvis-aligned) và PA-MPJPE (Procrustes-aligned) giữa ma trận tọa độ dự đoán và thực tế.
      operationId: evaluate_metrics
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateMetricsRequest'
      responses:
        '200':
          description: Tính toán thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculateMetricsResponse'
        '400':
          description: Dữ liệu đầu vào không hợp lệ hoặc kích thước ma trận không khớp
        '404':
          description: Không tìm thấy job_id hoặc file kết quả tương ứng

components:
  schemas:
    JobCreateRequest:
      type: object
      required:
        - video_id
      properties:
        video_id:
          type: string
          description: ID của video được trả về từ API `/api/v1/video/upload`
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        static_cam:
          type: boolean
          default: true
          description: Video quay từ camera tĩnh hay di chuyển
        use_dpvo:
          type: boolean
          default: true
          description: Sử dụng mô-đun DPVO cho phân tích chuyển động camera

    CalculateMetricsRequest:
      type: object
      required:
        - target_j3d
      properties:
        pred_j3d:
          type: array
          description: Tọa độ khớp 3D dự đoán dạng ma trận (F, J, 3) hoặc (J, 3). Tùy chọn nếu cung cấp job_id.
          items:
            type: array
            items:
              type: number
        target_j3d:
          type: array
          description: Tọa độ khớp 3D thực tế (Ground-Truth) dạng ma trận (F, J, 3) hoặc (J, 3).
          items:
            type: array
            items:
              type: number
        job_id:
          type: string
          description: (Tùy chọn) ID tác vụ đã hoàn thành để tự động nạp kết quả dự đoán.
          example: "550e8400-e29b-41d4-a716-446655440000"
        pelvis_idxs:
          type: array
          items:
            type: integer
          default: [1, 2]
          description: Chỉ số các khớp pelvis dùng để căn chỉnh gốc tọa độ.
        unit:
          type: string
          enum: [mm, m]
          default: "mm"
          description: Đơn vị đo sai số đầu ra.

    CalculateMetricsResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        unit:
          type: string
          example: "mm"
        num_frames:
          type: integer
          example: 100
        num_joints:
          type: integer
          example: 24
        mpjpe:
          type: object
          properties:
            mean:
              type: number
              example: 45.21
            min:
              type: number
              example: 20.15
            max:
              type: number
              example: 89.30
            std:
              type: number
              example: 12.40
            per_frame:
              type: array
              items:
                type: number
        pa_mpjpe:
          type: object
          properties:
            mean:
              type: number
              example: 32.14
            min:
              type: number
              example: 15.02
            max:
              type: number
              example: 65.10
            std:
              type: number
              example: 8.55
            per_frame:
              type: array
              items:
                type: number

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Cung cấp API Key để xác thực các request.